Modify for sqlite
