+++ .env.sample
@@ -1,4 +1,5 @@
-PORT="8008"
-LOG_DIR="./log"
+PORT="8080"
+LOG_DIR="./logs"
 LOG_FILE="log.csv"
-
+DB_DIR="./db"
+DB_FILE="logs.db"
