+++ .gitignore
@@ -4,6 +4,7 @@
 /temp
 /tmp
 /logs
+/db
 
 # Created by https://www.toptal.com/developers/gitignore/api/visualstudiocode,macos,windows,linux,vim,netbeans,go
 # Edit at https://www.toptal.com/developers/gitignore?templates=visualstudiocode,macos,windows,linux,vim,netbeans,go
