+++ test_commands/postError.sh
@@ -1,0 +1,2 @@
+curl -X POST -H "Content-Type: application/json" -d '{"message":"Error", "level":"error"}' localhost:8080/api/v1/log
+
